pyproject.toml
[tool.poetry]
name = "test-suite"
version = "0.1.0"
description = ""
authors = ["srinivas <srinivas@vananam.com>"]
maintainers = ["Amiya <amiya@vananam.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pytest = "8.1.1"
requests = "2.31.0"
bs4 = "0.0.2"
python-dotenv = "1.0.1"
icecream = "2.1.3"
jsonschema = "4.21.1"
faker = "24.11.0"
asttokens = "2.4.1"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
colorama = "0.4.6"
exceptiongroup = "1.2.1"
executing = "2.0.1"
idna = "3.7"
iniconfig = "2.0.0"
jsonschema-specifications = "2023.12.1"
packaging = "24.0"
pluggy = "1.4.0"
pygments = "2.17.2"
python-dateutil = "2.9.0.post0"
referencing = "0.34.0"
rpds-py = "0.18.0"
six = "1.16.0"
soupsieve = "2.5"
tomli = "2.0.1"
urllib3 = "2.2.1"
openpyxl = "^3.1.2"
selenium = "^4.19.0"
mysql-connector-python = "^8.3.0"
webdriver-manager = "^4.0.1"
pytest-html = "^4.1.1"
curlify = "^2.2.1"
pytest-order = "^1.2.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
